Add library
