+++ package.json
@@ -92,6 +92,7 @@
     "@hookform/error-message": "2.0.1",
     "@island.is/regulations-tools": "0.7.14",
     "@keyv/redis": "2.6.1",
+    "@livechat/widget-react": "1.3.2",
     "@nestjs/apollo": "10.1.0",
     "@nestjs/axios": "2.0.0",
     "@nestjs/bull": "0.6.3",
